rename Python files for PEP https://peps.python.org/pep-0008/#package-and-module-names
