+++ scrapers/GangBangCreampie/GangBangCreampie.yml
@@ -6,28 +6,28 @@ sceneByURL:
       - gangbangcreampie.com/en/video
     script:
       - python
-      - GangBangCreampie.py
+      - gangbang_creampie.py
       - gangbangcreampie
       - scene-by-url
 sceneByFragment:
     action: script
     script:
       - python
-      - GangBangCreampie.py
+      - gangbang_creampie.py
       - gangbangcreampie
       - scene-by-fragment
 sceneByName:
     action: script
     script:
       - python
-      - GangBangCreampie.py
+      - gangbang_creampie.py
       - gangbangcreampie
       - scene-by-name
 sceneByQueryFragment:
     action: script
     script:
       - python
-      - GangBangCreampie.py
+      - gangbang_creampie.py
       - gangbangcreampie
       - scene-by-query-fragment
 galleryByURL:
@@ -36,7 +36,7 @@ galleryByURL:
       - gangbangcreampie.com/en/photo/
     script:
       - python
-      - GangBangCreampie.py
+      - gangbang_creampie.py
       - gangbangcreampie
       - gallery-by-url
 # Last Updated January 30, 2025
