+++ scrapers/GenderXFilms/genderx_films.py
@@ -1,13 +1,11 @@
 """
-Stash scraper for Evil Angel (Network) that uses the Algolia API Python client
+Stash scraper for GenderX Films that uses the Algolia API Python client
 """
 import json
-import re
 import sys
 from typing import Any
-from urllib.parse import urlparse
 
-from AlgoliaAPI.AlgoliaAPI import (
+from scrapers.AlgoliaAPI.algolia_api import (
     ScrapedGallery,
     ScrapedMovie,
     gallery_from_fragment,
@@ -18,8 +16,7 @@ from AlgoliaAPI.AlgoliaAPI import (
     performer_search,
     scene_from_fragment,
     scene_from_url,
-    scene_search,
-    site_from_url
+    scene_search
 )
 
 from py_common import log
@@ -27,42 +24,19 @@ from py_common.types import ScrapedScene
 from py_common.util import scraper_args
 
 channel_name_map = {
-    "AnalPlaytime": "Anal Acrobats",
-    "Anal Trixxx": "AnalTriXXX",
-    "Jonni Darkko ": "Jonni Darkko XXX",    # trailing space is in the API
-    "Le Wood": "LeWood",
-    "Secret Crush ": "Secret Crush",    # trailing space is in the API
 }
 """
 This map just contains overrides when using a channel name as the studio
 """
 
 serie_name_map = {
-    "TransPlaytime": "TS Playground",
-    "XXXmailed": "Blackmailed",
 }
 """
 Each serie_name requiring a map/override should have a key-value here
 """
 
 site_map = {
-    "christophclarkonline": "Christoph Clark Online",
-    "christophsbignaturaltits": "Christoph's Big Natural Tits",
-    "gapingangels": "Gaping Angels",
-    "iloveblackshemales": "I Love Black Shemales",
-    "jakemalone": "Jake Malone",
-    "johnleslie": "John Leslie",
-    "lexingtonsteele": "Lexington Steele",
-    "nachovidalhardcore": "Nacho Vidal Hardcore",
-    "pansexualx": "PansexualX",
-    "pantypops": "Panty Pops",
-    "povblowjobs": "POV Blowjobs",
-    "roccosiffredi": "Rocco Siffredi",
-    "sheplayswithhercock": "She Plays With Her Cock",
-    "strapattackers": "Strap Attackers",
-    "tittycreampies": "Titty Creampies",
-    "transgressivexxx": "TransgressiveXXX",
-    "tsfactor": "TS Factor",
+    "genderxfilms": "GenderXFilms",
 }
 """
 Each site found in the logic should have a key-value here
@@ -91,67 +65,14 @@ def determine_studio(api_object: dict[str, Any]) -> str | None:
         return site_map.get(site_match, site_match)
     if serie_name in [
         *serie_name_map,
-        "PansexualX",
     ]:
         log.debug(f"matched serie_name '{serie_name}'")
         return serie_name_map.get(serie_name, serie_name)
-    if main_channel_name in [
-        *channel_name_map,
-        "Buttman",
-        "Cock Choking Sluts",
-        "Euro Angels",
-        "Transsexual Angel",
-        "TransgressiveXXX",
-    ]:
+    if main_channel_name:
         log.debug(f"matched main_channel_name '{main_channel_name}'")
+        # most scenes have the studio name as the main channel name
         return channel_name_map.get(main_channel_name, main_channel_name)
-    if director_match := next(
-        (item for item in [
-            "Joey Silvera",
-            "Mike Adriano",
-        ] if item in [c.get("name") for c in api_object.get("channels", [])]),
-        None
-    ):
-        log.debug(f"matched director_match '{director_match}'")
-        return director_match
-    if movie_desc := api_object.get("movie_desc"):
-        if "BAM Visions" in movie_desc:
-            log.debug("matched 'BAM Visions' in movie_desc")
-            return "BAM Visions"
-    log.debug("Did not match any studio override logic")
     return None
-
-
-def fix_ts_trans_find_replace(text: str) -> str | None:
-    """
-    At some point in time, there was a mass find-replace performed that replaced
-    all occurrences of "TS" or "ts" with "Trans".
-
-    The problem with this is that it replaced every match naively, resulting in
-    these examples:
-    - tits -> tiTrans
-    - hits -> hiTrans
-
-    This regex sub should undo those changes, but leave the intended change:
-    - TS -> Trans
-    """
-    if text:
-        return re.sub(r"(?<=[a-z])Trans", "ts", text)
-    return None
-
-
-def fix_url(_url: str) -> str:
-    """
-    Replaces the host part of the URL if criteria matched
-    """
-    if _url:
-        site = site_from_url(_url)
-        # if the site does not have a real/working domain
-        if site in [
-            "lexingtonsteele",
-        ]:
-            return urlparse(_url)._replace(netloc="www.evilangel.com").geturl()
-    return _url
 
 
 def postprocess_scene(scene: ScrapedScene, api_scene: dict[str, Any]) -> ScrapedScene:
@@ -160,15 +81,6 @@ def postprocess_scene(scene: ScrapedScene, api_scene: dict[str, Any]) -> Scraped
     """
     if studio_override := determine_studio(api_scene):
         scene["studio"] = { "name": studio_override }
-
-    if details := scene.get("details"):
-        scene["details"] = fix_ts_trans_find_replace(details)
-
-    if _url := scene.get("url"):
-        scene["url"] = fix_url(_url)
-
-    if urls := scene.get("urls"):
-        scene["urls"] = [fix_url(url) for url in urls]
 
     return scene
 
@@ -180,9 +92,6 @@ def postprocess_movie(movie: ScrapedMovie, api_movie: dict[str, Any]) -> Scraped
     if studio_override := determine_studio(api_movie):
         movie["studio"] = { "name": studio_override }
 
-    if synopsis := movie.get("synopsis"):
-        movie["synopsis"] = fix_ts_trans_find_replace(synopsis)
-
     return movie
 
 
@@ -192,9 +101,6 @@ def postprocess_gallery(gallery: ScrapedGallery, api_gallery: dict[str, Any]) ->
     """
     if studio_override := determine_studio(api_gallery):
         gallery["studio"] = { "name": studio_override }
-
-    if details := gallery.get("details"):
-        gallery["details"] = fix_ts_trans_find_replace(details)
 
     return gallery
