+++ scrapers/RoccoSiffredi/rocco_siffredi.py
@@ -5,7 +5,7 @@ import json
 import sys
 from typing import Any
 
-from AlgoliaAPI.AlgoliaAPI import (
+from scrapers.AlgoliaAPI.algolia_api import (
     gallery_from_fragment,
     gallery_from_url,
     movie_from_url,
